+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added `infrastructure.server.services.sftp.lockdown.auto` to activity
 - Billing tiers now have a `hard_cap` boolean for free tiers
+- Servers now have a features struct
 
 ### Changed